Create Container interface, add DiceContainer as implementation
